Fix column type

+++ db/migrate/20191027134936_create_action_rules.rb
@@ -3,8 +3,8 @@ class CreateActionRules < ActiveRecord::Migration[6.0]
     create_table :action_rules do |t|
       t.references :action_chain, null: false, foreign_key: true
       t.integer :motion_id
-      t.integer :move_x
-      t.integer :move_y
+      t.float :move_x
+      t.float :move_y
       t.integer :next, null: true
       t.references :search, null: true, foreign_key: true
       t.integer :target_value, null: true
